requirements.txt
requests-2.27.1
terminaltables-3.1.10
python-dotenv-0.20.0
